+++ R/groupr.R
@@ -1,5 +1,6 @@
 #' @import purrr vctrs
-#' @importFrom rlang syms abort enexprs
+#' @importFrom rlang syms abort enexprs := .data
 #' @importFrom tibble tibble as_tibble
 #' @importFrom dplyr group_data group_vars ungroup group_by group_cols
+#' @importFrom stats setNames
 NULL